Fix dll compilation issue for Visual Studio using makefile
The compilation sanity checks in dll.c are expecting _WINDLL to be
defined. In the Visual Studio project this definition is inherited
but is not present when building using the makefile. As _WINDLL is
referenced no where in the project other than the sanity checks it
seems safe to define it via the makefile.

This resolves #5.